feat(payment): INT-857 Pull request feedback from David C
