+++ src/customer/strategies/googlepay/googlepay-customer-mock.ts
@@ -20,7 +20,7 @@ export enum Mode {
     Incomplete,
 }
 
-export function getBraintreeCustomerInitializeOptions(mode: Mode = Mode.Full): CustomerInitializeOptions {
+export function getCustomerInitializeOptions(mode: Mode = Mode.Full): CustomerInitializeOptions {
     const methodId = { methodId: 'googlepaybraintree' };
     const undefinedMethodId = { methodId: undefined };
     const container = { container: 'googlePayCheckoutButton' };
@@ -40,30 +40,6 @@ export function getBraintreeCustomerInitializeOptions(mode: Mode = Mode.Full): C
         }
         default: {
             return { ...methodId, ...googlepayBraintree };
-        }
-     }
-}
-
-export function getStripeCustomerInitializeOptions(mode: Mode = Mode.Full): CustomerInitializeOptions {
-    const methodId = { methodId: 'googlepaystripe' };
-    const undefinedMethodId = { methodId: undefined };
-    const container = { container: 'googlePayCheckoutButton' };
-    const invalidContainer = { container: 'invalid_container' };
-    const googlepayStripe = { googlepaystripe: { ...container } };
-    const googlepayStripeWithInvalidContainer = { googlepaystripe: { ...invalidContainer } };
-
-    switch (mode) {
-        case Mode.Incomplete: {
-            return { ...methodId };
-        }
-        case Mode.UndefinedMethodId: {
-            return { ...undefinedMethodId, ...googlepayStripe };
-        }
-        case Mode.InvalidContainer: {
-            return { ...methodId, ...googlepayStripeWithInvalidContainer };
-        }
-        default: {
-            return { ...methodId, ...googlepayStripe };
         }
      }
 }
