+++ src/customer/strategies/googlepay/index.ts
@@ -1,3 +1,2 @@
-export { default as GooglePayBraintreeCustomerStrategy } from './googlepay-braintree-customer-strategy';
+export { default as GooglePayCustomerStrategy } from './googlepay-customer-strategy';
 export { default as GooglePayCustomerInitializeOptions } from './googlepay-customer-initialize-options';
-export { default as GooglePayStripeCustomerStrategy } from './googlepay-stripe-customer-strategy';
